+++ .github/workflows/pytest.yml
@@ -5,13 +5,13 @@ on:
     # Action will run when any changes to these paths are pushed or pr'ed to master
     branches: [ master ]
     paths:
-      - uli_init/**
+      - polybinder/**
       - environment-cpu.yml
       - .github/workflows/pytest.yml
   pull_request:
     branches: [ master ]
     paths:
-      - uli_init/**
+      - polybinder/**
       - environment-cpu.yml
       - .github/workflows/pytest.yml
   # Allows workflow to be manually triggered
