+++ polybinder/library/__init__.py
@@ -1,0 +1,3 @@
+from polybinder.library.systems import SYSTEM_DIR
+from polybinder.library.compounds import COMPOUND_DIR
+from polybinder.library.forcefields import FF_DIR
