+++ polybinder/tests/test_simulations.py
@@ -1,4 +1,4 @@
-from uli_init import simulate
+from polybinder import simulate
 from base_test import BaseTest
 
 import pytest
