--- uli_init/library/__init__.py
@@ -1,3 +1,0 @@
-from uli_init.library.systems import SYSTEM_DIR
-from uli_init.library.compounds import COMPOUND_DIR
-from uli_init.library.forcefields import FF_DIR
