--- uli_init/library/compounds/peek_meta.mol2
@@ -1,85 +1,0 @@
-@<TRIPOS>MOLECULE
-RES
-36 38 1 0 1
-SMALL
-NO_CHARGES
-@<TRIPOS>CRYSIN
-   20.3514     9.8199    10.2942    90.0000    90.0000    90.0000  1  1
-@<TRIPOS>ATOM
-       1 C            0.8166    -0.9093    -0.0294 C             1 RES     
-       2 C           -0.4575    -1.3974    -0.3378 C             1 RES     
-       3 C           -1.5528    -0.5283    -0.3635 C             1 RES     
-       4 C           -1.3747     0.8297    -0.0808 C             1 RES     
-       5 C           -0.1013     1.3196     0.2278 C             1 RES     
-       6 C            0.9978     0.4526     0.2546 C             1 RES     
-       7 O            2.2609     0.9521     0.5633 O             1 RES     
-       8 H            1.6600    -1.5878    -0.0113 H             1 RES     
-       9 H           -0.5959    -2.4483    -0.5567 H             1 RES     
-      10 H           -2.2227     1.5020    -0.1008 H             1 RES     
-      11 H            0.0319     2.3716     0.4460 H             1 RES     
-      12 C           -3.9785     0.5964     0.7408 C             1 RES     
-      13 C           -5.1404     1.1537     1.2851 C             1 RES     
-      14 C           -6.3925     0.6307     0.9470 C             1 RES     
-      15 C           -6.4836    -0.4501     0.0642 C             1 RES     
-      16 C           -5.3228    -1.0087    -0.4811 C             1 RES     
-      17 C           -4.0674    -0.4879    -0.1450 C             1 RES     
-      18 O           -2.9186    -1.0526    -0.6943 O             1 RES     
-      19 H           -3.0125     1.0063     1.0070 H             1 RES     
-      20 H           -5.0703     1.9902     1.9684 H             1 RES     
-      21 H           -7.2910     1.0624     1.3685 H             1 RES     
-      22 H           -5.3982    -1.8453    -1.1640 H             1 RES     
-      23 C           -8.7173    -0.3960     1.9430 C             1 RES     
-      24 C           -9.7551    -0.3457     2.8792 C             1 RES     
-      25 C          -11.0131    -0.8629     2.5594 C             1 RES     
-      26 C          -11.2358    -1.4315     1.3025 C             1 RES     
-      27 C          -10.2011    -1.4842     0.3628 C             1 RES     
-      28 C           -8.9317    -0.9656     0.6772 C             1 RES     
-      29 C           -7.8270    -1.0107    -0.2981 C             1 RES     
-      30 O           -7.9884    -1.5150    -1.4411 O             1 RES     
-      31 H           -7.7478     0.0093     2.2063 H             1 RES     
-      32 H           -9.5837     0.0945     3.8530 H             1 RES     
-      33 H          -12.2107    -1.8318     1.0560 H             1 RES     
-      34 H          -10.3935    -1.9287    -0.6056 H             1 RES     
-      35 H            3.1407     0.3024     0.5929 H             1 RES     
-      36 H          -11.8237    -0.8227     3.2929 H             1 RES     
-@<TRIPOS>BOND
-       1        2        1 1
-       2        6        1 1
-       3        8        1 1
-       4        3        2 1
-       5        9        2 1
-       6        4        3 1
-       7       18        3 1
-       8        5        6 1
-       9        7        6 1
-      10        4        5 1
-      11       11        5 1
-      12       35        7 1
-      13       10        4 1
-      14       13       12 1
-      15       17       12 1
-      16       19       12 1
-      17       14       13 1
-      18       20       13 1
-      19       15       14 1
-      20       21       14 1
-      21       16       17 1
-      22       18       17 1
-      23       15       16 1
-      24       22       16 1
-      25       29       15 1
-      26       24       23 1
-      27       28       23 1
-      28       31       23 1
-      29       25       24 1
-      30       32       24 1
-      31       26       25 1
-      32       36       25 1
-      33       27       28 1
-      34       29       28 1
-      35       26       27 1
-      36       34       27 1
-      37       30       29 1
-      38       33       26 1
-@<TRIPOS>SUBSTRUCTURE
-       1 RES             1 RESIDUE    0 **** ROOT      0
