--- uli_init/library/compounds/peek_para.mol2
@@ -1,85 +1,0 @@
-@<TRIPOS>MOLECULE
-RES
-36 38 1 0 1
-SMALL
-NO_CHARGES
-@<TRIPOS>CRYSIN
-   18.2122    10.3506    12.8976    90.0000    90.0000    90.0000  1  1
-@<TRIPOS>ATOM
-       1 C            0.7676    -0.9514     0.0144 C             1 RES     
-       2 C           -0.5395    -1.3846     0.2608 C             1 RES     
-       3 C           -1.5891    -0.4610     0.2908 C             1 RES     
-       4 C           -1.3324     0.8966     0.0745 C             1 RES     
-       5 C           -0.0259     1.3316    -0.1719 C             1 RES     
-       6 C            1.0276     0.4100    -0.2028 C             1 RES     
-       7 O            2.3243     0.8551    -0.4492 O             1 RES     
-       8 H            1.5753    -1.6719    -0.0074 H             1 RES     
-       9 H           -0.7387    -2.4352     0.4282 H             1 RES     
-      10 H           -2.1450     1.6111     0.0979 H             1 RES     
-      11 H            0.1681     2.3835    -0.3388 H             1 RES     
-      12 C           -3.3403     0.8278     2.1989 C             1 RES     
-      13 C           -4.1725     1.4642     3.1259 C             1 RES     
-      14 C           -5.4722     0.9939     3.3384 C             1 RES     
-      15 C           -5.9408    -0.1132     2.6241 C             1 RES     
-      16 C           -5.1101    -0.7510     1.6967 C             1 RES     
-      17 C           -3.8082    -0.2831     1.4811 C             1 RES     
-      18 O           -2.9904    -0.9264     0.5550 O             1 RES     
-      19 H           -2.3353     1.1972     2.0394 H             1 RES     
-      20 H           -3.8102     2.3212     3.6790 H             1 RES     
-      21 H           -6.9468    -0.4769     2.7889 H             1 RES     
-      22 H           -5.4771    -1.6077     1.1460 H             1 RES     
-      23 C           -7.4997    -0.4792     4.8353 C             1 RES     
-      24 C           -8.4557    -1.2302     5.5268 C             1 RES     
-      25 C           -9.2948    -0.6109     6.4567 C             1 RES     
-      26 C           -9.1797     0.7608     6.6969 C             1 RES     
-      27 C           -8.2253     1.5161     6.0078 C             1 RES     
-      28 C           -7.3767     0.9000     5.0701 C             1 RES     
-      29 C           -6.3637     1.6771     4.3326 C             1 RES     
-      30 O           -6.2293     2.9154     4.5214 O             1 RES     
-      31 H           -6.8567    -0.9742     4.1177 H             1 RES     
-      32 H           -8.5457    -2.2928     5.3415 H             1 RES     
-      33 H           -9.8303     1.2393     7.4174 H             1 RES     
-      34 H           -8.1516     2.5778     6.2080 H             1 RES     
-      35 H            3.1702     0.1620    -0.4802 H             1 RES     
-      36 H          -10.0420    -1.2000     6.9967 H             1 RES     
-@<TRIPOS>BOND
-       1        2        1 1
-       2        6        1 1
-       3        8        1 1
-       4        3        2 1
-       5        9        2 1
-       6        4        3 1
-       7       18        3 1
-       8        5        6 1
-       9        7        6 1
-      10        4        5 1
-      11       11        5 1
-      12       35        7 1
-      13       10        4 1
-      14       13       12 1
-      15       17       12 1
-      16       19       12 1
-      17       14       13 1
-      18       20       13 1
-      19       15       14 1
-      20       29       14 1
-      21       16       17 1
-      22       18       17 1
-      23       15       16 1
-      24       22       16 1
-      25       21       15 1
-      26       24       23 1
-      27       28       23 1
-      28       31       23 1
-      29       25       24 1
-      30       32       24 1
-      31       26       25 1
-      32       36       25 1
-      33       27       28 1
-      34       29       28 1
-      35       26       27 1
-      36       34       27 1
-      37       30       29 1
-      38       33       26 1
-@<TRIPOS>SUBSTRUCTURE
-       1 RES             1 RESIDUE    0 **** ROOT      0
