--- uli_init/library/compounds/pekk_meta.mol2
@@ -1,87 +1,0 @@
-@<TRIPOS>MOLECULE
-RES
-37 39 1 0 1
-SMALL
-NO_CHARGES
-@<TRIPOS>CRYSIN
-   20.3925    11.3669     8.8932    90.0000    90.0000    90.0000  1  1
-@<TRIPOS>ATOM
-       1 C            0.7676    -0.9487     0.0742 C             1 RES     
-       2 C           -0.5395    -1.3655     0.3472 C             1 RES     
-       3 C           -1.5891    -0.4418     0.3192 C             1 RES     
-       4 C           -1.3324     0.8995     0.0181 C             1 RES     
-       5 C           -0.0259     1.3182    -0.2552 C             1 RES     
-       6 C            1.0276     0.3964    -0.2281 C             1 RES     
-       7 O            2.3243     0.8252    -0.5020 O             1 RES     
-       8 H            1.5753    -1.6691     0.0976 H             1 RES     
-       9 H           -0.7387    -2.4035     0.5803 H             1 RES     
-      10 H           -2.1450     1.6141    -0.0035 H             1 RES     
-      11 H            0.1681     2.3575    -0.4878 H             1 RES     
-      12 C            5.8062    -0.5928     0.0237 C             1 RES     
-      13 C            4.7235     0.2924     0.0066 C             1 RES     
-      14 C            3.4841    -0.1260    -0.4847 C             1 RES     
-      15 C            3.3250    -1.4303    -0.9599 C             1 RES     
-      16 C            4.4048    -2.3192    -0.9449 C             1 RES     
-      17 C            5.6558    -1.9057    -0.4518 C             1 RES     
-      18 C            6.8074    -2.8260    -0.4267 C             1 RES     
-      19 O            6.7037    -4.0094    -0.8462 O             1 RES     
-      20 H            6.7605    -0.2531     0.4078 H             1 RES     
-      21 H            4.8456     1.3030     0.3744 H             1 RES     
-      22 H            2.3645    -1.7532    -1.3402 H             1 RES     
-      23 H            4.2615    -3.3258    -1.3174 H             1 RES     
-      24 C           10.5942    -1.4832     1.0979 C             1 RES     
-      25 C            9.4038    -0.9889     1.6408 C             1 RES     
-      26 C            8.1727    -1.4259     1.1447 C             1 RES     
-      27 C            8.1298    -2.3581     0.1046 C             1 RES     
-      28 C            9.3175    -2.8553    -0.4416 C             1 RES     
-      29 C           10.5608    -2.4203     0.0523 C             1 RES     
-      30 C           11.8249    -2.9313    -0.5087 C             1 RES     
-      31 O           11.8275    -3.7722    -1.4467 O             1 RES     
-      32 H           11.5403    -1.1344     1.4935 H             1 RES     
-      33 H            9.4359    -0.2669     2.4464 H             1 RES     
-      34 H            7.2525    -1.0424     1.5660 H             1 RES     
-      35 H            9.2632    -3.5770    -1.2471 H             1 RES     
-      36 H           -2.6111    -0.7685     0.5328 H             1 RES     
-      37 H           12.7814    -2.5802    -0.1103 H             1 RES     
-@<TRIPOS>BOND
-       1        2        1 1
-       2        6        1 1
-       3        8        1 1
-       4        3        2 1
-       5        9        2 1
-       6        4        3 1
-       7       36        3 1
-       8        5        6 1
-       9        7        6 1
-      10        4        5 1
-      11       11        5 1
-      12       14        7 1
-      13       10        4 1
-      14       13       12 1
-      15       17       12 1
-      16       20       12 1
-      17       14       13 1
-      18       21       13 1
-      19       15       14 1
-      20       16       17 1
-      21       18       17 1
-      22       15       16 1
-      23       23       16 1
-      24       19       18 1
-      25       27       18 1
-      26       22       15 1
-      27       25       24 1
-      28       29       24 1
-      29       32       24 1
-      30       26       25 1
-      31       33       25 1
-      32       27       26 1
-      33       34       26 1
-      34       28       29 1
-      35       30       29 1
-      36       27       28 1
-      37       35       28 1
-      38       31       30 1
-      39       37       30 1
-@<TRIPOS>SUBSTRUCTURE
-       1 RES             1 RESIDUE    0 **** ROOT      0
