--- uli_init/library/compounds/pekk_para.mol2
@@ -1,87 +1,0 @@
-@<TRIPOS>MOLECULE
-RES
-37 39 1 0 1
-SMALL
-NO_CHARGES
-@<TRIPOS>CRYSIN
-   19.0836    10.5428    12.5944    90.0000    90.0000    90.0000  1  1
-@<TRIPOS>ATOM
-       1 C            0.7676    -0.6830    -0.6626 C             1 RES     
-       2 C           -0.5395    -1.1634    -0.7947 C             1 RES     
-       3 C           -1.5891    -0.5316    -0.1203 C             1 RES     
-       4 C           -1.3324     0.5813     0.6867 C             1 RES     
-       5 C           -0.0259     1.0632     0.8200 C             1 RES     
-       6 C            1.0276     0.4333     0.1465 C             1 RES     
-       7 O            2.3243     0.9223     0.2870 O             1 RES     
-       8 H            1.5753    -1.1770    -1.1875 H             1 RES     
-       9 H           -0.7387    -2.0247    -1.4192 H             1 RES     
-      10 H           -2.1450     1.0700     1.2084 H             1 RES     
-      11 H            0.1681     1.9250     1.4458 H             1 RES     
-      12 C            5.4453     0.4649    -1.8251 C             1 RES     
-      13 C            4.3633     1.0613    -1.1696 C             1 RES     
-      14 C            3.4841     0.2803    -0.4150 C             1 RES     
-      15 C            3.6852    -1.0988    -0.3144 C             1 RES     
-      16 C            4.7659    -1.6999    -0.9680 C             1 RES     
-      17 C            5.6555    -0.9204    -1.7297 C             1 RES     
-      18 C            6.8010    -1.5333    -2.4268 C             1 RES     
-      19 O            7.0157    -2.7728    -2.3602 O             1 RES     
-      20 H            6.1178     1.0842    -2.4061 H             1 RES     
-      21 H            4.2065     2.1295    -1.2469 H             1 RES     
-      22 H            3.0035    -1.7027     0.2705 H             1 RES     
-      23 H            4.9048    -2.7700    -0.8778 H             1 RES     
-      24 C            9.9621    -0.2060    -4.0565 C             1 RES     
-      25 C            9.0907    -1.0022    -3.3063 C             1 RES     
-      26 C            7.7336    -0.6778    -3.2320 C             1 RES     
-      27 C            7.2453     0.4435    -3.9079 C             1 RES     
-      28 C            8.1128     1.2428    -4.6594 C             1 RES     
-      29 C            9.4804     0.9229    -4.7391 C             1 RES     
-      30 C           10.4125     1.7516    -5.5254 C             1 RES     
-      31 O           10.0106     2.7700    -6.1485 O             1 RES     
-      32 H           11.0114    -0.4709    -4.1040 H             1 RES     
-      33 H            9.4679    -1.8709    -2.7824 H             1 RES     
-      34 H            6.1938     0.6937    -3.8496 H             1 RES     
-      35 H            7.7152     2.1075    -5.1759 H             1 RES     
-      36 H           -2.6111    -0.9079    -0.2241 H             1 RES     
-      37 H           11.4725     1.4855    -5.5746 H             1 RES     
-@<TRIPOS>BOND
-       1        2        1 1
-       2        6        1 1
-       3        8        1 1
-       4        3        2 1
-       5        9        2 1
-       6        4        3 1
-       7       36        3 1
-       8        5        6 1
-       9        7        6 1
-      10        4        5 1
-      11       11        5 1
-      12       14        7 1
-      13       10        4 1
-      14       13       12 1
-      15       17       12 1
-      16       20       12 1
-      17       14       13 1
-      18       21       13 1
-      19       15       14 1
-      20       16       17 1
-      21       18       17 1
-      22       15       16 1
-      23       23       16 1
-      24       19       18 1
-      25       26       18 1
-      26       22       15 1
-      27       25       24 1
-      28       29       24 1
-      29       32       24 1
-      30       26       25 1
-      31       33       25 1
-      32       27       26 1
-      33       28       29 1
-      34       30       29 1
-      35       27       28 1
-      36       35       28 1
-      37       31       30 1
-      38       37       30 1
-      39       34       27 1
-@<TRIPOS>SUBSTRUCTURE
-       1 RES             1 RESIDUE    0 **** ROOT      0
