--- uli_init/library/forcefields/M-M.txt
@@ -1,101 +1,0 @@
-1.000000000000000048e-04 6.104845921251013579e+02 8.257475943015540452e+02
-7.010000000000000953e-02 5.526822605239925679e+02 8.257475943015532494e+02
-1.401000000000000023e-01 4.948799289228838916e+02 8.257475943015532494e+02
-2.101000000000000090e-01 4.370775973217751016e+02 8.257475943015537041e+02
-2.801000000000000156e-01 3.792752657206663685e+02 8.257475943015537041e+02
-3.501000000000000223e-01 3.214729341195575785e+02 8.257475943015537041e+02
-4.201000000000000290e-01 2.636706025184488453e+02 8.257475943015537041e+02
-4.901000000000000356e-01 2.058682709173400553e+02 8.257475943015538178e+02
-5.601000000000000423e-01 1.480659393162312938e+02 8.257475943015540452e+02
-6.301000000000001044e-01 9.026360771512247538e+01 8.257475943015537041e+02
-7.001000000000000556e-01 3.246127611401375646e+01 5.145587163401982025e+02
-7.701000000000000068e-01 1.822538742749472718e+01 1.375148421180963112e+02
-8.401000000000000689e-01 1.320919821748026912e+01 2.559248802820283331e+01
-9.101000000000001311e-01 1.464243910354633016e+01 3.145188419020808368e+00
-9.801000000000000822e-01 1.276887183881735588e+01 3.982359366860391958e+01
-1.050100000000000033e+00 9.067135989941780494e+00 4.108417378218274507e+01
-1.120100000000000096e+00 7.017087509311770610e+00 2.290539374505665648e+01
-1.190100000000000158e+00 5.860380865633848302e+00 1.264315326662519823e+01
-1.260100000000000220e+00 5.247046051984242787e+00 1.221503509889690875e+01
-1.330100000000000060e+00 4.150275951788280793e+00 1.120815090746742193e+01
-1.400100000000000122e+00 3.677904924938803610e+00 6.133659387454758694e+00
-1.470100000000000184e+00 3.291563637544614451e+00 5.734648832477457603e+00
-1.540100000000000025e+00 2.875054088391959528e+00 7.736146526524015243e+00
-1.610100000000000087e+00 2.208503123831252246e+00 7.753903739570391807e+00
-1.680100000000000149e+00 1.789507564852104604e+00 4.220180067603644503e+00
-1.750100000000000211e+00 1.617677914366741909e+00 7.565895078785371819e-01
-1.820100000000000273e+00 1.683585033749109394e+00 1.740447935769168897e+00
-1.890100000000000113e+00 1.374015203359058246e+00 4.060250822430148787e+00
-1.960100000000000176e+00 1.115149918608888457e+00 5.425081545712397535e+00
-2.030100000000000460e+00 6.145037869593225066e-01 2.680329434173155612e+00
-2.100100000000000300e+00 7.399037978246466585e-01 -9.275740160475625176e-03
-2.170100000000000584e+00 6.158023905817890942e-01 2.924804065366514827e+00
-2.240100000000000424e+00 3.304312286733345205e-01 2.828875516434278037e+00
-2.310100000000000264e+00 2.197598182809901157e-01 2.545876377390595913e+00
-2.380100000000000549e+00 -2.599146416134897122e-02 1.991228929010681714e+00
-2.450100000000000389e+00 -5.901223178050537393e-02 1.906587282315488352e+00
-2.520100000000000673e+00 -2.929136836855173454e-01 1.999845831207864855e+00
-2.590100000000000513e+00 -3.389906481496064439e-01 4.704632213019100395e-01
-2.660100000000000353e+00 -3.587785346677847542e-01 3.087809355790405719e-01
-2.730100000000000637e+00 -3.822199791306721273e-01 5.420289363015247108e-01
-2.800100000000000477e+00 -4.346625857499982204e-01 6.089040795186866362e-01
-2.870100000000000318e+00 -4.674665502632882608e-01 1.897358616965731937e-01
-2.940100000000000602e+00 -4.612256063875184720e-01 -3.810738791200719844e-02
-3.010100000000000442e+00 -4.621315159556072527e-01 1.083861734857395343e+00
-3.080100000000000282e+00 -6.129662492675538310e-01 7.559140795422235604e-01
-3.150100000000000566e+00 -5.679594870915185556e-01 -2.790341477446492413e-01
-3.220100000000000406e+00 -5.739014685833029317e-01 -4.929630835439798298e-01
-3.290100000000000691e+00 -4.989446553953613694e-01 1.771442137193403321e-01
-3.360100000000000531e+00 -5.987016585040105809e-01 1.275507923024288948e+00
-3.430100000000000371e+00 -6.775157646187618399e-01 -1.050763473699089928e-01
-3.500100000000000655e+00 -5.839909698722233200e-01 -5.676505742910770147e-01
-3.570100000000000495e+00 -5.980446842180110467e-01 1.415582883689044946e+00
-3.640100000000000779e+00 -7.821725735886896347e-01 6.049190578761084458e-01
-3.710100000000000620e+00 -6.827333523206662313e-01 -8.233661058421753998e-01
-3.780100000000000460e+00 -6.669013187707850676e-01 -3.975577364427503868e-01
-3.850100000000000744e+00 -6.270752692186811750e-01 -6.649784925434807370e-01
-3.920100000000000584e+00 -5.738043298146977556e-01 -7.365453262845338411e-01
-3.990100000000000424e+00 -5.239589235388464328e-01 -6.516614904469066882e-01
-4.060100000000000264e+00 -4.825717211521308170e-01 -1.032093397115603217e+00
-4.130100000000000549e+00 -3.794658479426619713e-01 -4.201017886113742605e-01
-4.200099999999999945e+00 -4.237574707465384116e-01 -2.209046794368781097e-01
-4.270100000000000229e+00 -3.485391928214990331e-01 -8.896447067798461950e-01
-4.340100000000000513e+00 -2.992072117973599288e-01 -7.403455264365962352e-01
-4.410099999999999909e+00 -2.448908191203755524e-01 -9.480268868358390222e-01
-4.480100000000000193e+00 -1.664834476403424524e-01 -4.590589389967185130e-01
-4.550100000000000477e+00 -1.806225676608349540e-01 1.372983435251683981e-01
-4.620099999999999874e+00 -1.857052157338660303e-01 1.374894676789336356e-01
-4.690100000000000158e+00 -1.998710931358856635e-01 -5.127462482965464141e-02
-4.760100000000000442e+00 -1.785267682577143800e-01 -6.171209143031456401e-01
-4.830099999999999838e+00 -1.134741651334452706e-01 -5.896888865415288450e-01
-4.900100000000000122e+00 -9.597032414190033445e-02 2.194007592607113630e-02
-4.970100000000000406e+00 -1.165457757630952301e-01 -6.024492861088467932e-02
-5.040100000000000691e+00 -8.753603413637647823e-02 -2.342943065035174577e-01
-5.110100000000000087e+00 -8.374457285260278327e-02 -1.734111533130064386e-02
-5.180100000000000371e+00 -8.510827798999438809e-02 -2.169516690945049245e-02
-5.250100000000000655e+00 -8.070724948527971399e-02 -2.856375219973461554e-01
-5.320100000000000051e+00 -4.511902491036592217e-02 -3.127128754923369502e-01
-5.390100000000000335e+00 -3.692744691635253818e-02 7.582950685035347305e-02
-5.460100000000000620e+00 -5.573515586941540922e-02 -5.987948347843380370e-02
-5.530100000000000016e+00 -2.854431922937180482e-02 -3.895646611674173365e-01
-5.600100000000000300e+00 -1.196103305976979440e-03 -1.258001798779144753e-01
-5.670100000000000584e+00 -1.093229404646377592e-02 1.051685051546304428e-01
-5.740099999999999980e+00 -1.591969402762524396e-02 1.245177628692118246e-02
-5.810100000000000264e+00 -1.267554272663274154e-02 -9.781614687646640605e-02
-5.880100000000000549e+00 -2.225433464919945176e-03 1.225112325216790032e-01
-5.950099999999999945e+00 -2.982711527966780407e-02 2.140181406167503964e-01
-6.020100000000000229e+00 -3.218797315126500219e-02 -1.967277921397524165e-01
-6.090100000000000513e+00 -2.285224380102461421e-03 -1.906032402701184347e-01
-6.160099999999999909e+00 -5.503519513448418450e-03 1.426160243848595133e-01
-6.230100000000000193e+00 -2.225146779398279664e-02 2.057868061882256183e-01
-6.300100000000000477e+00 -3.431367237980000734e-02 7.782758614817669729e-02
-6.370100000000000762e+00 -3.314732985472753496e-02 -1.094510223679616284e-01
-6.440100000000000158e+00 -1.899052924828537742e-02 -1.117177978781121739e-01
-6.510100000000000442e+00 -1.750683815179182937e-02 -1.615275585902789246e-01
-6.580100000000000726e+00 3.623328954353675826e-03 -1.948452718855604171e-01
-6.650100000000000122e+00 9.771499912186629513e-03 -5.710031752201642080e-02
-6.720100000000000406e+00 1.161737340743597587e-02 5.683633748810137248e-02
-6.790100000000000691e+00 1.814412663852436620e-03 8.164448505755973751e-02
-6.860100000000000087e+00 1.871454993776117159e-04 1.283601755939406397e-02
-6.930100000000000371e+00 1.737020553726735940e-05 1.336753566982940618e-03
-7.000100000000000655e+00 0.000000000000000000e+00 2.481457933895336670e-04
