--- uli_init/library/forcefields/P-P.txt
@@ -1,101 +1,0 @@
-1.000000000000000048e-04 1.111776428132763641e+02 1.453673991381173778e+02
-7.010000000000000953e-02 1.010019248736081465e+02 1.453673991381173778e+02
-1.401000000000000023e-01 9.082620693393992894e+01 1.453673991381173778e+02
-2.101000000000000090e-01 8.065048899427171136e+01 1.453673991381172925e+02
-2.801000000000000156e-01 7.047477105460350799e+01 1.453673991381172925e+02
-3.501000000000000223e-01 6.029905311493529041e+01 1.453673991381173778e+02
-4.201000000000000290e-01 5.012333517526707283e+01 1.453673991381173209e+02
-4.901000000000000356e-01 3.994761723559886235e+01 1.453673991381172925e+02
-5.601000000000000423e-01 2.977189929593064832e+01 -5.075305255429286681e-14
-6.301000000000001044e-01 3.994761723559886946e+01 4.320498301819959863e+01
-7.001000000000000556e-01 2.372320167338270380e+01 1.609892987617328117e+02
-7.701000000000000068e-01 1.740911540895627496e+01 9.373463485203259893e+01
-8.401000000000000689e-01 1.060035279409813924e+01 4.933453351722142344e+01
-9.101000000000001311e-01 1.050228071654527540e+01 2.425146861497649464e+00
-9.801000000000000822e-01 1.026083223348846829e+01 1.402541946592870659e+01
-1.050100000000000033e+00 8.538721991315256332e+00 2.985417545890064517e+01
-1.120100000000000096e+00 6.081247669242377363e+00 2.176144427517648339e+01
-1.190100000000000158e+00 5.492119792790548516e+00 1.001773119769468856e+01
-1.260100000000000220e+00 4.678765301565120716e+00 1.233003188776370784e+01
-1.330100000000000060e+00 3.765915328503629311e+00 1.451838917679387464e+01
-1.400100000000000122e+00 2.646190816813978053e+00 7.943117323330220358e+00
-1.470100000000000184e+00 2.653878903237398390e+00 -9.433864711230008915e-01
-1.540100000000000025e+00 2.778264922771198187e+00 7.912366147877903932e-01
-1.610100000000000087e+00 2.543105777167107728e+00 8.179673587640831300e+00
-1.680100000000000149e+00 1.633110620501481591e+00 8.237361093340876295e+00
-1.750100000000000211e+00 1.389875224099385020e+00 -1.091576655745220448e+00
-1.820100000000000273e+00 1.785931352305812458e+00 -1.125943795687878168e+00
-1.890100000000000113e+00 1.547507355495687964e+00 4.918898091257382532e+00
-1.960100000000000176e+00 1.097285619529778877e+00 3.869405787889766302e+00
-2.030100000000000460e+00 1.005790545191120655e+00 1.511769997230836715e+00
-2.100100000000000300e+00 8.856378199174617105e-01 2.455414002416555774e-01
-2.170100000000000584e+00 9.714147491572888704e-01 2.303229745765632774e+00
-2.240100000000000424e+00 5.631856555102731088e-01 4.176811553177676828e+00
-2.310100000000000264e+00 3.866611317124141189e-01 -7.548513984777888153e-01
-2.380100000000000549e+00 6.688648512971635540e-01 1.651110486402869448e-01
-2.450100000000000389e+00 3.635455849027739439e-01 4.167748679935594502e+00
-2.520100000000000673e+00 8.538003610618027928e-02 2.929036329216444301e+00
-2.590100000000000513e+00 -4.651950118752829627e-02 1.867555738006990262e+00
-2.660100000000000353e+00 -1.760777672147984252e-01 -3.886092934072288840e-01
-2.730100000000000637e+00 7.885799889483754641e-03 -2.339032712291135296e-01
-2.800100000000000477e+00 -1.433313092427225266e-01 2.534331999059646900e+00
-2.870100000000000318e+00 -3.469206799788668860e-01 1.456648188662480470e+00
-2.940100000000000602e+00 -3.472620556554698035e-01 3.572703434692288127e-01
-3.010100000000000442e+00 -3.969385280645589265e-01 1.484389686635548822e+00
-3.080100000000000282e+00 -5.550766117844466718e-01 1.815018525189234344e+00
-3.150100000000000566e+00 -6.510411215910517635e-01 6.285174520770883122e-01
-3.220100000000000406e+00 -6.430690550752390378e-01 -3.660265947637432182e-01
-3.290100000000000691e+00 -5.997973983241277063e-01 7.152350151446907889e-01
-3.360100000000000531e+00 -7.432019571954957549e-01 7.161965768168123692e-01
-3.430100000000000371e+00 -7.000649190784814513e-01 -4.017824524564450051e-01
-3.500100000000000655e+00 -6.869524138515934464e-01 5.641288052731460168e-01
-3.570100000000000495e+00 -7.790429518167218959e-01 1.407793655079278006e+00
-3.640100000000000779e+00 -8.840435255626923894e-01 2.045111071395330216e-01
-3.710100000000000620e+00 -8.076745068162565211e-01 -1.095163725903571805e+00
-3.780100000000000460e+00 -7.307206039361923144e-01 -7.865880613704776536e-01
-3.850100000000000744e+00 -6.975521782243896407e-01 -1.022262300501024790e-01
-3.920100000000000584e+00 -7.164089317291779668e-01 -5.063449472132457441e-01
-3.990100000000000424e+00 -6.266638856145352321e-01 -6.377060041945253888e-01
-4.060100000000000264e+00 -6.271300911419444013e-01 -1.822637456599543807e-02
-4.130100000000000549e+00 -6.241121931752958707e-01 4.813466438125761637e-02
-4.200099999999999945e+00 -6.338689441553204684e-01 -7.766694451032356561e-01
-4.270100000000000229e+00 -5.153784708608428744e-01 -1.397047090236052336e+00
-4.340100000000000513e+00 -4.382823515222731281e-01 -1.385966128376558082e+00
-4.410099999999999909e+00 -3.213432128881247296e-01 -1.044299812847231612e+00
-4.480100000000000193e+00 -2.920803777236606891e-01 -6.207847475585955843e-01
-4.550100000000000477e+00 -2.344333482299213456e-01 -5.438790345186431674e-01
-4.620099999999999874e+00 -2.159373128910506423e-01 -7.698254157694907063e-01
-4.690100000000000158e+00 -1.266577900221926289e-01 -1.903208437371739226e-01
-4.760100000000000442e+00 -1.892923947678462893e-01 3.075510209567017816e-01
-4.830099999999999838e+00 -1.697149329561308817e-01 -3.731915371424401329e-01
-4.900100000000000122e+00 -1.370455795679046684e-01 -4.117646007748793080e-01
-4.970100000000000406e+00 -1.120678888476477758e-01 -2.835244416923368149e-01
-5.040100000000000691e+00 -9.735215773097750935e-02 -1.675520073288949974e-01
-5.110100000000000087e+00 -8.861060782160247451e-02 -1.338622733873967019e-01
-5.180100000000000371e+00 -7.861143945674196887e-02 -5.103384268220790831e-01
-5.250100000000000655e+00 -1.716322806651138927e-02 -2.135276464634791316e-01
-5.320100000000000051e+00 -4.871756895185488934e-02 2.055575131153774548e-01
-5.390100000000000335e+00 -4.594127990266423656e-02 -1.905073139289879258e-01
-5.460100000000000620e+00 -2.204654500179657750e-02 -3.273135082557280806e-01
-5.530100000000000016e+00 -1.173887468623067629e-04 -2.733409133478912345e-01
-5.600100000000000300e+00 1.622118286690819866e-02 -7.488647889670217717e-02
-5.670100000000000584e+00 1.036671829867599759e-02 1.237991864861526098e-01
-5.740099999999999980e+00 -1.110703241153168609e-03 2.103108982766769275e-01
-5.810100000000000264e+00 -1.907680746005877698e-02 2.795548318148107780e-01
-5.880100000000000549e+00 -4.024837969522668257e-02 1.902679049856750891e-03
-5.950099999999999945e+00 -1.934318252703872212e-02 -5.133574968044973298e-03
-6.020100000000000229e+00 -3.952967919970038629e-02 4.956775197322622895e-02
-6.090100000000000513e+00 -2.628266780329039515e-02 -2.088528001546234580e-01
-6.160099999999999909e+00 -1.029028717805309870e-02 -2.430189285684267453e-01
-6.230100000000000193e+00 7.739982196289354847e-03 -4.776107523082469969e-02
-6.300100000000000477e+00 -3.603736645737640187e-03 7.895624317605318132e-02
-6.370100000000000762e+00 -3.313891848358093209e-03 3.000404358903498159e-02
-6.440100000000000158e+00 -7.804302748202538061e-03 -4.734589811292619875e-03
-6.510100000000000442e+00 -2.651049274777126340e-03 8.526140037677004679e-02
-6.580100000000000726e+00 -1.974089880095034402e-02 1.025554258946553832e-01
-6.650100000000000122e+00 -1.700880890002888154e-02 -5.336608196566472451e-02
-6.720100000000000406e+00 -1.226964732575728211e-02 -1.185567978233597608e-01
-6.790100000000000691e+00 -4.108572047585126491e-04 -8.777045798756659412e-02
-6.860100000000000087e+00 1.821679250204310543e-05 -3.032538510064461141e-03
-6.930100000000000371e+00 1.369818665051197641e-05 1.301199464431650155e-04
-7.000100000000000655e+00 0.000000000000000000e+00 1.956883807215996330e-04
